restructure folder for the 3rd fucking time
