+++ app/page.tsx
@@ -1,13 +1,13 @@
 import '../styles/globals.css'
 
-import Desktop from '../src/components/Desktop'
-import TaskBar from '../src/components/Desktop/TaskBar'
+import Desktop from './Desktop'
+import TaskBar from './Desktop/TaskBar'
 
 // data
-import getAllShortcuts from './fetches/getAllShortcuts'
+import useGetAllShortcuts from './hooks/useGetAllShortcuts'
 
 export default function App() {
-  const allShortCuts = getAllShortcuts('type')
+  const allShortCuts = useGetAllShortcuts('type')
   return (
     <div className='dark h-screen w-screen max-w-[100vw] max-h-[100vh]'>
       <Desktop allShortcuts={allShortCuts} />
