+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "date-fns": "^2.29.3",
+    "gray-matter": "^4.0.3",
     "next": "13.0.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
