--- src/components/Desktop/Window/WindowContent/index.tsx
@@ -1,7 +1,0 @@
-export default () => {
-  return (
-    <div className='bg-gray border-0 border-t-0 border-solid border-gray'>
-      Window Content
-    </div>
-  )
-}
